Combine key-mapper with xbindkeys
There's a bug in KDE that makes it regularly not recognize multi-key
combos sent from key-mapper. Instead KDE passess them directly to the app
in focus, which often garbles whatever I'm working on (on top of not
performing the operation I wanted to perform).

Apparently this is true only for multi-key combos with KDE running on
top of X11. I did not find the issue to be present when running on
wayland, but that was very short testing, so maybe I just missed the
occurences.

Anyway, unfortunately this means this configs is 50-80% incompatible
with Wayland, as all the functionality of xbindkeys doesn't work out of
the box and possibly some of those KEY_F?? / XF86Launch? keys that KDE
catches on X11 might not work on wayland either.